Add LoadingIndicator component and integrate it into EditSatus and EditScreenLayout
